Revert "Add translations"
