+++ .idea/.name
@@ -1,1 +1,1 @@
-Simple Bible App
+CUTView 2
