+++ app/build.gradle.kts
@@ -4,11 +4,11 @@ plugins {
 }
 
 android {
-    namespace = "com.example.simplebibleapp"
+    namespace = "com.example.cutview2"
     compileSdk = 34
 
     defaultConfig {
-        applicationId = "com.example.simplebibleapp"
+        applicationId = "com.example.cutview2"
         minSdk = 24
         targetSdk = 34
         versionCode = 1
