+++ app/src/main/AndroidManifest.xml
@@ -7,7 +7,7 @@
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
         android:icon="@mipmap/ic_launcher"
-        android:label="@string/Simple_Bible_App"
+        android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.CUTView2"
@@ -15,7 +15,7 @@
         <activity
             android:name=".MainActivity"
             android:exported="true"
-            android:label="@string/Simple_Bible_App"
+            android:label="@string/app_name"
             android:theme="@style/Theme.CUTView2">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
