+++ app/src/main/java/com/example/cutview2/MainUiState.kt
@@ -1,0 +1,7 @@
+package com.example.cutview2
+
+data class MainUiState(
+    var bookname : String = "",
+    var chapter : Int = 0,
+    var zoom : Float = 1f
+)
