+++ app/src/main/java/com/example/cutview2/MainViewModelFactory.kt
@@ -1,0 +1,16 @@
+package com.example.cutview2
+
+import androidx.datastore.core.DataStore
+import androidx.datastore.preferences.core.Preferences
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+
+class MainViewModelFactory(val readdata: ReadBibleData, val dataStore: DataStore<Preferences>) : ViewModelProvider.Factory {
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
+        if (modelClass.isAssignableFrom(MainViewModel::class.java)) {
+            @Suppress("UNCHECKED_CAST")
+            return MainViewModel(readdata, dataStore) as T
+        }
+        throw IllegalArgumentException("Unable to construct viewmodel")
+    }
+}
