+++ app/src/main/java/com/example/cutview2/ReadDummyData.kt
@@ -1,0 +1,15 @@
+package com.example.cutview2
+
+class ReadDummyData : ReadBibleData{
+    override fun getBooknamesList(): List<String> {
+        TODO("Not yet implemented")
+    }
+
+    override fun getChapterCount(bookname: String): Int {
+        TODO("Not yet implemented")
+    }
+
+    override fun getChapterFromBook(bookname: String, chapter: Int): List<String> {
+        TODO("Not yet implemented")
+    }
+}
