+++ app/src/main/java/com/example/cutview2/UserPreferences.kt
@@ -1,0 +1,6 @@
+package com.example.cutview2
+
+data class UserPreferences(
+    val bookname: String,
+    val chapter: Int
+)
