--- app/src/main/java/com/example/simplebibleapp/UserPreferences.kt
@@ -1,6 +1,0 @@
-package com.example.simplebibleapp
-
-data class UserPreferences(
-    val bookname: String,
-    val chapter: Int
-)
