--- app/src/main/java/com/example/simplebibleapp/dataClasses/BookDetails.kt
@@ -1,8 +1,0 @@
-package com.example.simplebibleapp.dataClasses
-
-class BookDetails(
-    val bookName: String,
-    val bookIndex : Int,
-    val chapterCount : Int,
-    val language : String
-)
