--- app/src/main/java/com/example/simplebibleapp/dataClasses/LocalisedPromptIds.kt
@@ -1,6 +1,0 @@
-package com.example.simplebibleapp.dataClasses
-
-class LocalisedPromptIds(
-    val bookPrompt: Int,
-    val chapterPrompt: Int
-)
