+++ app/src/main/res/values/arrays.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
 
-    <array name="booknames_tcn">
+    <array name="booknames">
         <item>創世紀</item>
         <item>出埃及記</item>
         <item>利未記</item>
@@ -68,74 +68,6 @@
         <item>約翰三書</item>
         <item>猶大書</item>
         <item>啟示錄</item>
-    </array>
-    <array name="booknames_en">
-        <item>Genesis</item>
-        <item>Exodus</item>
-        <item>Leviticus</item>
-        <item>Numbers</item>
-        <item>Deuteronomy</item>
-        <item>Joshua</item>
-        <item>Judges</item>
-        <item>Ruth</item>
-        <item>1 Samuel</item>
-        <item>2 Samuel</item>
-        <item>1 Kings</item>
-        <item>2 Kings</item>
-        <item>1 Chronicles</item>
-        <item>2 Chronicles</item>
-        <item>Ezra</item>
-        <item>Nehemiah</item>
-        <item>Esther</item>
-        <item>Job</item>
-        <item>Psalm</item>
-        <item>Proverbs</item>
-        <item>Ecclesiastes</item>
-        <item>Song of Solomon</item>
-        <item>Isaiah</item>
-        <item>Jeremiah</item>
-        <item>Lamentations</item>
-        <item>Ezekiel</item>
-        <item>Daniel</item>
-        <item>Hosea</item>
-        <item>Joel</item>
-        <item>Amos</item>
-        <item>Obadiah</item>
-        <item>Jonah</item>
-        <item>Micah</item>
-        <item>Nahum</item>
-        <item>Habakkuk</item>
-        <item>Zephaniah</item>
-        <item>Haggai</item>
-        <item>Zechariah</item>
-        <item>Malachi</item>
-        <item>Matthew</item>
-        <item>Mark</item>
-        <item>Luke</item>
-        <item>John</item>
-        <item>Acts</item>
-        <item>Romans</item>
-        <item>1 Corinthians</item>
-        <item>2 Corinthians</item>
-        <item>Galatians</item>
-        <item>Ephesians</item>
-        <item>Philippians</item>
-        <item>Colossians</item>
-        <item>1 Thessalonians</item>
-        <item>2 Thessalonians</item>
-        <item>1 Timothy</item>
-        <item>2 Timothy</item>
-        <item>Titus</item>
-        <item>Philemon</item>
-        <item>Hebrews</item>
-        <item>James</item>
-        <item>1 Peter</item>
-        <item>2 Peter</item>
-        <item>1 John</item>
-        <item>2 John</item>
-        <item>3 John</item>
-        <item>Jude</item>
-        <item>Revelation</item>
     </array>
     <array name="chapternumbers">
         <item>50</item>
