+++ app/src/main/res/values/strings.xml
@@ -1,7 +1,5 @@
 <resources>
-    <string name="Simple_Bible_App">Simple_Bible_App</string>
-    <string name="book_prompt_tcn">請選擇書卷</string>
-    <string name="chapter_prompt_tcn">請選擇章</string>
-    <string name="book_prompt_en">Book</string>
-    <string name="chapter_prompt_en">Chapter</string>
+    <string name="app_name">CUTView 2</string>
+    <string name="book_prompt">請選擇書卷</string>
+    <string name="chapter_prompt">請選擇章</string>
 </resources>
